server/requirements.txt
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
PyMySQL==1.0.2
Pillow==9.3.0
